Try to refactor

+++ MelonModManager/Core/WindowColor.cs
@@ -4,12 +4,12 @@ namespace MelonModManager.Core
 {
     public static class WindowColor
     {
-        public static Color ModListLayout = new Color32((byte) 18, (byte) 18, (byte) 18, (byte) 255);
-        public static Color ModInfoLayout = new Color32((byte) 30, (byte) 30, (byte) 30, (byte) 255);
-        public static Color RoundLayout = new Color32((byte) 45, (byte) 49, (byte) 53, (byte) 255);
-        public static Color Invisible = new Color32((byte) 0, (byte) 0, (byte) 0, (byte) 0);
-        public static Color Translucent = new Color32((byte) 0, (byte) 0, (byte) 0, (byte) 150);
-        public static Color ButtonColor = new Color32((byte)195, (byte) 143, (byte) 255, (byte) 255);
+        public static Color ModListLayout = new Color32(18, 18, 18, 255);
+        public static Color ModInfoLayout = new Color32(30, 30, 30, 255);
+        public static Color RoundLayout = new Color32(45, 49, 53, 255);
+        public static Color Invisible = new Color32(0, 0, 0, 0);
+        public static Color Translucent = new Color32(0, 0, 0, 150);
+        public static Color ButtonColor = new Color32(195, 143, 255, 255);
         
         public static string SelectText = "#bb86fc";
         public static string NotSelectText = "#969696";
